+++ server/src/events/mod.rs
@@ -1,4 +1,7 @@
-use crate::{events::interaction::handle_tame_pet, state_ext::StateExt, Server};
+use crate::{
+    events::interaction::handle_tame_pet, persistence::PersistedComponents, state_ext::StateExt,
+    Server,
+};
 use common::event::{EventBus, ServerEvent};
 use common_base::span;
 use entity_creation::{
@@ -6,10 +9,10 @@ use entity_creation::{
     handle_initialize_character, handle_loaded_character_data, handle_shockwave, handle_shoot,
 };
 use entity_manipulation::{
-    handle_aura, handle_bonk, handle_buff, handle_combo_change, handle_delete, handle_destroy,
-    handle_energy_change, handle_entity_attacked_hook, handle_explosion, handle_health_change,
-    handle_knockback, handle_land_on_ground, handle_parry, handle_poise, handle_respawn,
-    handle_teleport_to,
+    handle_aura, handle_bonk, handle_buff, handle_change_ability, handle_combo_change,
+    handle_delete, handle_destroy, handle_energy_change, handle_entity_attacked_hook,
+    handle_explosion, handle_health_change, handle_knockback, handle_land_on_ground, handle_parry,
+    handle_poise, handle_respawn, handle_teleport_to,
 };
 use group_manip::handle_group;
 use information::handle_site_info;
@@ -129,6 +132,17 @@ impl Server {
                     character_id,
                 } => handle_initialize_character(self, entity, character_id),
                 ServerEvent::UpdateCharacterData { entity, components } => {
+                    let (body, stats, skill_set, inventory, waypoint, pets, active_abilities) =
+                        components;
+                    let components = PersistedComponents {
+                        body,
+                        stats,
+                        skill_set,
+                        inventory,
+                        waypoint,
+                        pets,
+                        active_abilities,
+                    };
                     handle_loaded_character_data(self, entity, components);
                 },
                 ServerEvent::ExitIngame { entity } => {
@@ -233,6 +247,12 @@ impl Server {
                 ServerEvent::EntityAttackedHook { entity } => {
                     handle_entity_attacked_hook(self, entity)
                 },
+                ServerEvent::ChangeAbility {
+                    entity,
+                    slot,
+                    auxiliary_key,
+                    new_ability,
+                } => handle_change_ability(self, entity, slot, auxiliary_key, new_ability),
             }
         }