+++ voxygen/src/hud/util.rs
@@ -365,52 +365,41 @@ pub fn ability_description(ability_id: &str) -> (&str, &str) {
         // Debug stick
         "common.abilities.debug.possess" => (
             "Possessing Arrow",
-            "\n\
-            Shoots a poisonous arrow.\n\
-            Lets you control your target.",
+            "Shoots a poisonous arrow. Lets you control your target.",
         ),
         // Sword
         "common.abilities.sword.spin" => (
             "Whirlwind",
-            "\n\
-            Move forward while spinning with your sword.",
+            "Move forward while spinning with your sword.",
         ),
         // Axe
         "common.abilities.axe.leap" => (
             "Axe Jump",
-            "\n\
-            A jump with the slashing leap to position of cursor.",
+            "A jump with the slashing leap to position of cursor.",
         ),
         // Hammer
         "common.abilities.hammer.leap" => (
             "Smash of Doom",
-            "\n\
-            An AOE attack with knockback.\n\
-            Leaps to position of cursor.",
+            "An AOE attack with knockback. Leaps to position of cursor.",
         ),
         // Bow
         "common.abilities.bow.shotgun" => (
             "Burst",
-            "\n\
-            Launches a burst of arrows",
+            "Launches a burst of arrows",
         ),
         // Staff
         "common.abilities.staff.fireshockwave" => (
             "Ring of Fire",
-            "\n\
-            Ignites the ground with fiery shockwave.",
+            "Ignites the ground with fiery shockwave.",
         ),
         // Sceptre
         "common.abilities.sceptre.wardingaura" => (
             "Thorn Bulwark",
-            "\n\
-            Protects you and your group with thorns\n\
-            for a short amount of time.",
+            "Protects you and your group with thorns for a short amount of time.",
         ),
         _ => (
             "Ability has no title",
-            "\n\
-            Ability has no description."
+            "Ability has no description."
         ),
     }
 }