+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Piercing damage now ignores an amount of protection equal to damage value
 - Slashing damage now reduces target's energy by an amount equal to damage dealt to target post-mitigation
 - Crushing damage now does poise damage to a target equal to the amount mitigated by armor
+- UI to select abilities and assign to hotbar
+- Position of abilities on hotbar is now persisted through the server
 
 ### Changed